chore: clean up utils folder
