--- packages/api/src/utils/ApiError.ts
@@ -1,11 +1,0 @@
-export class ApiError extends Error {
-  statusCode: number
-
-  isOperational: boolean
-
-  constructor(statusCode: number, message: string, isOperational = true) {
-    super(message)
-    this.statusCode = statusCode
-    this.isOperational = isOperational
-  }
-}
